requirements.txt
torch
torchvision
transformers
timm
einops
clip-interrogator==0.6.0
onnxruntime
pillow
scikit-learn
